+++ internal/config/data/blacklist.txt
@@ -1,0 +1,86 @@
+.3gp
+.7z
+.a
+.aac
+.ac3
+.ai
+.apk
+.ass
+.avi
+.bin
+.bmp
+.class
+.cr2
+.cue
+.dat
+.deb
+.dll
+.dmg
+.dng
+.doc
+.docx
+.eps
+.exe
+.flac
+.flv
+.gif
+.gz
+.ico
+.img
+.ipa
+.iso
+.jar
+.jpg
+.jpeg
+.lib
+.m2v
+.m4a
+.m4v
+.mka
+.mkv
+.mov
+.mp3
+.mp4
+.mpeg
+.mpg
+.mpv
+.msi
+.nef
+.obj
+.ogv
+.orf
+.otf
+.pdf
+.pef
+.pem
+.pkg
+.png
+.ppk
+.ppt
+.pptx
+.psd
+.rar
+.raw
+.rm
+.rmvb
+.rpm
+.so
+.sr2
+.srt
+.ssa
+.sub
+.svg
+.swf
+.tar
+.tiff
+.ttf
+.vob
+.wav
+.webm
+.wmv
+.woff
+.woff2
+.xls
+.xlsx
+.xmp
+.zip
